fix: bound OpenClaw media upload inputs
